requirements.txt
openai>=1.0.0
python-dotenv
jupyter
pandas
numpy
matplotlib
seaborn
